+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 All notable changes to `laravel-pest-recorder` will be documented in this file.
 
 ## 1.0.1 - 2026-03-05
-* Ensure Browser tests directory exists
+* Ensure tests/Browser directory exists
 * Implement assertText -> assertSee method
 * Implement --visit=... option
 * Fix for clicking buttons without any id or test-id
 * Implement --acting-as=user option
+* Improve --viewport-size=... options and support fullscreen as a value
+* Implement --device=... option
 
 ## 1.0.0 - 2026-02-14
 * Initial release